Automated datatable tables.

Wrapper for DT::datatable for shorthand declaration of many layout arguments. Automates many formatting options.

Usage

robotable(
  d,
  title = NULL,
  subtitle = "",
  caption = NULL,
  rounding = 1,
  width = getOption("roboplot.width"),
  height = NULL,
  flag = "",
  unit = "",
  dateformat = NULL,
  pagelength = 10,
  info_text = NULL,
  heatmap = NULL,
  na_value = ""
)

Arguments

d

Data frame. Data to be created a table from.

title, subtitle

Characters. Labels for plot elements. Optionally, use set_title() for the title if you want to omit the title from the displayed plot, but include it for any downloads through the modebar.

caption

Function or character. Use set_caption().

rounding

Double. Controls the rounding of numeric columns.

height, width

Double. Height and width of the table. Default width is NULL for responsive plots, give a value for static table dimensions.

flag, unit

Character vectors. Use "+" for flag if you want to flag a numeric column for positive values. Unit is prepended to values in a numeric column. Name the units with column names from data 'd' if you want to target specific columns.

dateformat

Character. Controls how to format dates displayed on the table. Robotable tries to determine the proper format if left NULL.

pagelength

Double. Controls how many rows are displayed on the table. If data 'd' contains more rows than this, robotable() automatically adds navigation.

info_text

Character. Optional. If included, this text will be displayed with a popup when the info button in the table modebar is clicked.

heatmap

Function. Use set_heatmap(). Displays any numeric values as a heatmap.

na_value

Character. How NA values are displayed.

Value

A list of classes "datatable" and "htmlwidget"

Examples

# You can use roboplotr::robotable() to create html tables
#
energiantuonti |> robotable()
#> roboplotr message: Using the attribute "robonomist_title" for plot title.
#> roboplotr message: Using the attribute "source" for plot caption.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.

#
# No matter if you have any number date, character or numeric columns, robotable()
# handles the formatting automatically.
#
d <- energiantuonti |>
  dplyr::filter(Alue %in% c("Ruotsi","Kanada")) |>
  tidyr::unite(Tiedot, Alue, Suunta, sep = ", ") |>
  dplyr::arrange(Tiedot, time) |>
  tidyr::pivot_wider(names_from = Tiedot) |>
  dplyr::mutate(dplyr::across(where(is.numeric), ~ tidyr::replace_na(.x, 0))) |>
  dplyr::arrange(time)
d |> robotable()
#> roboplotr alert: Missing the title, using placeholder.
#> roboplotr alert: Missing the caption, using placeholder.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.

#
# You can use the pagelength parameter to control how many rows are shown at once,
# info_text to add any information you wish describe about the table, and
# heatmap parameter if you want to color-code any numeric values across the
# table. See [set_heatmap()] documentation for further examples.
#
d |>
 robotable(heatmap = set_heatmap(),
           pagelength = 50,
           info_text = "Testing the info button.")
#> roboplotr alert: Missing the title, using placeholder.
#> roboplotr alert: Missing the caption, using placeholder.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.