Set roboplot() pattern definitions

Set roboplot() pattern definitions

Usage

set_pattern(
  pattern = NULL,
  pattern_types = NULL,
  pattern_along = NULL,
  show_legend = TRUE,
  sep = ", "
)

Arguments

pattern

Symbol, string, or function resulting in symbol or string. Column from param 'd' of roboplot() to use for scatter plot linetype or bar plot pattern. Not supported for pie charts.

pattern_types

Named character vector. Pattern types for all traces. Names must correspond to values in the column referenced by param 'pattern' in param 'd' of roboplot(). See set_roboplot_options() parameter 'dashtypes' and 'patterns' for options, or use ".other" if all traces are of the same 'plot_type'.

pattern_along

Symbol, string, or function resulting in symbol or string. Column from param 'd' of roboplot() along which the series should be continuous along. If set, and the 'plot_mode' corresponding to all items in the given group as defined by 'pattern' is "line", roboplot() will try to fill the series in such a way as to create a continuous line.

show_legend

Logical, named if vector. If any pattern(s) will have their own legend entries. If FALSE, only the first item in param 'pattern' will be in roboplot() legend. If named, the must correspond to values in the column referenced by param 'pattern' in param 'd' of roboplot(). See set_roboplot_options() parameter 'dashtypes' and 'patterns' for options, or use ".other" if all traces are of the same 'plot_type'.

sep

Character. The separator of color and pattern in legend text. Default is ", ". Use NA if you want to omit the pattern label from legend, resulting only the roboplot() param 'color' being the label.

Examples

# You can use [set_pattern()] to just give the column name, but in that case
# you could just as well provide the column name by itself.
# Compare this:
energiantuonti |>
  dplyr::filter(Alue %in% c("Belgia","Kanada")) |>
  roboplot(
    Alue,
    pattern = Suunta
  )
#> roboplotr message: Using the attribute "robonomist_title" for plot title.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.
#> roboplotr message: Using the attribute "source" for plot caption.

#' # To this:
energiantuonti |>
  dplyr::filter(Alue %in% c("Belgia","Kanada")) |>
  roboplot(
    Alue,
    pattern = set_pattern(Suunta)
  )
#> roboplotr message: Using the attribute "robonomist_title" for plot title.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.
#> roboplotr message: Using the attribute "source" for plot caption.

# Suppose you have a series where some values are predictions and some are
# observations, you would have a gap in a line plot.
d <- energiantuonti |>
  dplyr::filter(Alue == "Belgia", Suunta == "Tuonti") |>
  dplyr::mutate(Sarjatyyppi = ifelse(
    lubridate::year(time) == max(lubridate::year(time)),
    "Ennuste",
    "Toteuma"
  ))

roboplot(d, Alue, pattern = Sarjatyyppi)
#> roboplotr message: Using the attribute "robonomist_title" for plot title.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.
#> roboplotr message: Using the attribute "source" for plot caption.

# Use the parameter 'pattern_along' to provide the column along which the plot
# should show a continous line.
roboplot(d, Alue, pattern = set_pattern(Sarjatyyppi, pattern_along = time))
#> roboplotr message: Using the attribute "robonomist_title" for plot title.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.
#> roboplotr message: Using the attribute "source" for plot caption.
#> roboplotr warning: roboplotr attempts to make continous line with different patterns along 'Sarjatyyppi' over 'time. Be sure to arrange the data in proper order. For complex data you are better off handling the data manipulation outside of roboplot().
#> Error in map(split_d, function(g) {    tracetype <- unique(g$roboplot.plot.type)    hoverlab <- case_when(tracetype == "pie" ~ "label", any(c("horizontal",         "horizontalfill", "horizontalstack") %in% g$roboplot.plot.type) &         is.null(pattern) & quo_name(color) != ticktypes$y ~ "text",         TRUE ~ "customdata")    hovertemplate <- roboplotr_hovertemplate(hovertext, lab = hoverlab,         ticktypes)    legend_rank <- mean(g$roboplot.legend.rank)    if (tracetype == "pie") {        g <- mutate(g, roboplot.bg.color = roboplotr_alter_color(.data$roboplot.trace.color,             "dark"), roboplot.tx.color = roboplotr_text_color_picker(roboplotr_alter_color(.data$roboplot.trace.color,             "dark")), roboplot.in.tx.color = roboplotr_text_color_picker(.data$roboplot.trace.color))    }    if (is.null(highlight)) {        if (!is.null(pattern_showlegend)) {            show.legend <- pattern_showlegend[as.character(unique(g[[as_name(pattern)]]))]        }        else {            show.legend <- trace_showlegend        }    }    else {        show.legend <- roboplotr_highlight_legend(highlight,             g)    }    roboplotr_check_param(legend_title, c("logical", "character"),         allow_null = F)    legendgrouptitle <- if (legend_title == F) {        NULL    }    else if (is.character(legend_title)) {        str_c("<b>", legend_title, "</b>")    }    else {        str_c("<b>", as_name(color), "</b>")    }    if (!is.null(error_bars)) {        if (quo_is_null(error_bars$error_x)) {            error_x <- NULL        }        else {            error_x <- list(array = error_bars$error_x, color = error_bars$xcolor)        }        if (quo_is_null(error_bars$error_y)) {            error_y <- NULL        }        else {            error_y <- list(array = error_bars$error_y, color = error_bars$ycolor)        }    }    else {        error_x <- NULL        error_y <- NULL    }    plotting_params <- list(color = color, customdata = if (any(c("horizontal",         "horizontalfill", "horizontalstack") %in% g$roboplot.plot.mode)) {        ~roboplot.horizontal.label    } else {        ~roboplot.plot.text    }, data = g, direction = "clockwise", error_x = error_x,         error_y = error_y, hoverlabel = list(bgcolor = ~roboplot.bg.color,             color = ~roboplot.tx.color), hovertemplate = hovertemplate,         insidetextfont = list(family = getOption("roboplot.font.main")$family,             size = getOption("roboplot.font.main")$size, color = ~roboplot.in.tx.color),         labels = color, legendgroup = color, legendrank = legend_rank,         line = ~list(width = roboplot.linewidth, dash = roboplot.dash,             smoothing = 0.5, shape = case_when(roboplot.plot.mode ==                 "step" ~ "hv", roboplot.plot.mode == "smooth" ~                 "spline", TRUE ~ "linear")), marker = list(colors = ~roboplot.trace.color,             line = getOption("roboplot.trace.border"), symbol = markers$symbol,             size = markers$size, pattern = ~list(shape = roboplot.pattern)),         mode = str_replace_all(unique(g$roboplot.plot.mode),             c(`^smooth$` = "line", `^line$` = "lines", `^step$` = "lines",                 `^scatter$` = "markers", `scatter\\+line` = "markers+lines")),         name = ~if (!is.null(legend_maxwidth)) {            str_trunc(as.character(roboplot.plot.text), legend_maxwidth)        } else {            roboplot.plot.text        }, offset = if ("horizontalfill" %in% g$roboplot.plot.mode) {            ~roboplot.bar.offset        } else {            NULL        }, offsetgroup = ~str_c(roboplot.pattern, roboplot.trace.color),         orientation = ifelse(any(c("horizontal", "horizontalfill",             "horizontalstack") %in% g$roboplot.plot.mode) & "bar" %in%             g$roboplot.plot.type, "h", "v"), rotation = rotation,         showlegend = show.legend, sort = F, text = if (any(c("horizontal",             "horizontalfill", "horizontalstack") %in% g$roboplot.plot.mode)) {            ~roboplot.horizontal.label        } else {            ~roboplot.plot.text        }, textinfo = "percent", textposition = case_when("horizontal" %in%             g$roboplot.plot.mode ~ "auto", tracetype == "pie" ~             "inside", TRUE ~ "none"), texttemplate = if (tracetype ==             "pie") {            NULL        } else if ("horizontal" %in% g$roboplot.plot.mode) {            "%{x:,.1f}"        } else {            NA        }, type = ~tracetype, legendgrouptitle = list(text = legendgrouptitle,             font = getOption("roboplot.font.main")[c("color",                 "family", "size")]), values = as.formula(str_c("~",             ticktypes$y)), width = if ("horizontalfill" %in%             g$roboplot.plot.mode) {            ~roboplot.bar.width        } else {            NULL        }, x = as.formula(str_c("~", ticktypes$x)), y = as.formula(str_c("~",             ifelse(!is.null(legend_maxwidth) & ticktypes$y !=                 "value", "roboplot.trunc", ticktypes$y))))    shared_params <- c("data", "text", "texttemplate", "hovertemplate",         "legendgroup", "showlegend", "type", "legendgrouptitle",         "customdata", "error_y", "error_x")    if (tracetype %in% "scatter" & any(c("line", "step", "smooth") %in%         g$roboplot.plot.mode)) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "line", "mode", "name", "color", "xhoverformat")]    }    else if (tracetype == "scatter" & "scatter+line" %in% g$roboplot.plot.mode) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "line", "mode", "name", "marker", "color", "xhoverformat")]    }    else if (tracetype == "scatter" & "scatter" %in% g$roboplot.plot.mode) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "mode", "name", "color", "marker", "xhoverformat")]    }    else if (tracetype == "bar" & any(c("horizontal", "horizontalfill",         "horizontalstack") %in% g$roboplot.plot.mode)) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "offsetgroup", "orientation", "offset", "width",             "color", "name", "textposition", "marker")]    }    else if (tracetype == "bar") {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "offsetgroup", "name", "color", "textposition",             "marker", "xhoverformat")]    }    else if (tracetype == "pie") {        plotting_params <- plotting_params[c(shared_params, "labels",             "textposition", "textinfo", "insidetextfont", "direction",             "rotation", "sort", "hoverlabel", "marker", "values")]    }    if (!is.null(secondary_yaxis)) {        if (length(unique(as.numeric(pull(g, {            {                secondary_yaxis            }        })))) > 1) {            stop("The column referred to by argument 'secondary_yaxis' must be plotted to one axis per trace. Try setting it to match the argument 'color'.",                 call. = F)        }        if (unique(as.numeric(pull(g, {            {                secondary_yaxis            }        }))) == 2) {            plotting_params$yaxis <- "y2"        }    }    plotting_params}): ℹ In index: 1.
#> Caused by error in `if (tracetype == "pie") ...`:
#> ! missing value where TRUE/FALSE needed

# Use the parameter 'show_legend' to omit all patterns expect the first one
# from the legend, and use the parameter 'sep' to control the separator
# between the roboplotr::roboplot() param 'color' and 'pattern' in the legend.
roboplot(d,
         Alue,
         pattern = set_pattern(
           Sarjatyyppi,
           pattern_along = time,
           show_legend = FALSE,
           sep = " - "
         ))
#> roboplotr message: Using the attribute "robonomist_title" for plot title.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.
#> roboplotr message: Using the attribute "source" for plot caption.
#> roboplotr warning: roboplotr attempts to make continous line with different patterns along 'Sarjatyyppi' over 'time. Be sure to arrange the data in proper order. For complex data you are better off handling the data manipulation outside of roboplot().
#> Error in map(split_d, function(g) {    tracetype <- unique(g$roboplot.plot.type)    hoverlab <- case_when(tracetype == "pie" ~ "label", any(c("horizontal",         "horizontalfill", "horizontalstack") %in% g$roboplot.plot.type) &         is.null(pattern) & quo_name(color) != ticktypes$y ~ "text",         TRUE ~ "customdata")    hovertemplate <- roboplotr_hovertemplate(hovertext, lab = hoverlab,         ticktypes)    legend_rank <- mean(g$roboplot.legend.rank)    if (tracetype == "pie") {        g <- mutate(g, roboplot.bg.color = roboplotr_alter_color(.data$roboplot.trace.color,             "dark"), roboplot.tx.color = roboplotr_text_color_picker(roboplotr_alter_color(.data$roboplot.trace.color,             "dark")), roboplot.in.tx.color = roboplotr_text_color_picker(.data$roboplot.trace.color))    }    if (is.null(highlight)) {        if (!is.null(pattern_showlegend)) {            show.legend <- pattern_showlegend[as.character(unique(g[[as_name(pattern)]]))]        }        else {            show.legend <- trace_showlegend        }    }    else {        show.legend <- roboplotr_highlight_legend(highlight,             g)    }    roboplotr_check_param(legend_title, c("logical", "character"),         allow_null = F)    legendgrouptitle <- if (legend_title == F) {        NULL    }    else if (is.character(legend_title)) {        str_c("<b>", legend_title, "</b>")    }    else {        str_c("<b>", as_name(color), "</b>")    }    if (!is.null(error_bars)) {        if (quo_is_null(error_bars$error_x)) {            error_x <- NULL        }        else {            error_x <- list(array = error_bars$error_x, color = error_bars$xcolor)        }        if (quo_is_null(error_bars$error_y)) {            error_y <- NULL        }        else {            error_y <- list(array = error_bars$error_y, color = error_bars$ycolor)        }    }    else {        error_x <- NULL        error_y <- NULL    }    plotting_params <- list(color = color, customdata = if (any(c("horizontal",         "horizontalfill", "horizontalstack") %in% g$roboplot.plot.mode)) {        ~roboplot.horizontal.label    } else {        ~roboplot.plot.text    }, data = g, direction = "clockwise", error_x = error_x,         error_y = error_y, hoverlabel = list(bgcolor = ~roboplot.bg.color,             color = ~roboplot.tx.color), hovertemplate = hovertemplate,         insidetextfont = list(family = getOption("roboplot.font.main")$family,             size = getOption("roboplot.font.main")$size, color = ~roboplot.in.tx.color),         labels = color, legendgroup = color, legendrank = legend_rank,         line = ~list(width = roboplot.linewidth, dash = roboplot.dash,             smoothing = 0.5, shape = case_when(roboplot.plot.mode ==                 "step" ~ "hv", roboplot.plot.mode == "smooth" ~                 "spline", TRUE ~ "linear")), marker = list(colors = ~roboplot.trace.color,             line = getOption("roboplot.trace.border"), symbol = markers$symbol,             size = markers$size, pattern = ~list(shape = roboplot.pattern)),         mode = str_replace_all(unique(g$roboplot.plot.mode),             c(`^smooth$` = "line", `^line$` = "lines", `^step$` = "lines",                 `^scatter$` = "markers", `scatter\\+line` = "markers+lines")),         name = ~if (!is.null(legend_maxwidth)) {            str_trunc(as.character(roboplot.plot.text), legend_maxwidth)        } else {            roboplot.plot.text        }, offset = if ("horizontalfill" %in% g$roboplot.plot.mode) {            ~roboplot.bar.offset        } else {            NULL        }, offsetgroup = ~str_c(roboplot.pattern, roboplot.trace.color),         orientation = ifelse(any(c("horizontal", "horizontalfill",             "horizontalstack") %in% g$roboplot.plot.mode) & "bar" %in%             g$roboplot.plot.type, "h", "v"), rotation = rotation,         showlegend = show.legend, sort = F, text = if (any(c("horizontal",             "horizontalfill", "horizontalstack") %in% g$roboplot.plot.mode)) {            ~roboplot.horizontal.label        } else {            ~roboplot.plot.text        }, textinfo = "percent", textposition = case_when("horizontal" %in%             g$roboplot.plot.mode ~ "auto", tracetype == "pie" ~             "inside", TRUE ~ "none"), texttemplate = if (tracetype ==             "pie") {            NULL        } else if ("horizontal" %in% g$roboplot.plot.mode) {            "%{x:,.1f}"        } else {            NA        }, type = ~tracetype, legendgrouptitle = list(text = legendgrouptitle,             font = getOption("roboplot.font.main")[c("color",                 "family", "size")]), values = as.formula(str_c("~",             ticktypes$y)), width = if ("horizontalfill" %in%             g$roboplot.plot.mode) {            ~roboplot.bar.width        } else {            NULL        }, x = as.formula(str_c("~", ticktypes$x)), y = as.formula(str_c("~",             ifelse(!is.null(legend_maxwidth) & ticktypes$y !=                 "value", "roboplot.trunc", ticktypes$y))))    shared_params <- c("data", "text", "texttemplate", "hovertemplate",         "legendgroup", "showlegend", "type", "legendgrouptitle",         "customdata", "error_y", "error_x")    if (tracetype %in% "scatter" & any(c("line", "step", "smooth") %in%         g$roboplot.plot.mode)) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "line", "mode", "name", "color", "xhoverformat")]    }    else if (tracetype == "scatter" & "scatter+line" %in% g$roboplot.plot.mode) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "line", "mode", "name", "marker", "color", "xhoverformat")]    }    else if (tracetype == "scatter" & "scatter" %in% g$roboplot.plot.mode) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "mode", "name", "color", "marker", "xhoverformat")]    }    else if (tracetype == "bar" & any(c("horizontal", "horizontalfill",         "horizontalstack") %in% g$roboplot.plot.mode)) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "offsetgroup", "orientation", "offset", "width",             "color", "name", "textposition", "marker")]    }    else if (tracetype == "bar") {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "offsetgroup", "name", "color", "textposition",             "marker", "xhoverformat")]    }    else if (tracetype == "pie") {        plotting_params <- plotting_params[c(shared_params, "labels",             "textposition", "textinfo", "insidetextfont", "direction",             "rotation", "sort", "hoverlabel", "marker", "values")]    }    if (!is.null(secondary_yaxis)) {        if (length(unique(as.numeric(pull(g, {            {                secondary_yaxis            }        })))) > 1) {            stop("The column referred to by argument 'secondary_yaxis' must be plotted to one axis per trace. Try setting it to match the argument 'color'.",                 call. = F)        }        if (unique(as.numeric(pull(g, {            {                secondary_yaxis            }        }))) == 2) {            plotting_params$yaxis <- "y2"        }    }    plotting_params}): ℹ In index: 1.
#> Caused by error in `if (tracetype == "pie") ...`:
#> ! missing value where TRUE/FALSE needed

# Or just use 'sep' = NA to omit the pattern from the legend labels.
roboplot(d, Alue,
           pattern =
             set_pattern(
               Sarjatyyppi,
               pattern_along = time,
               show_legend = FALSE,
               sep = NA
             ))
#> roboplotr message: Using the attribute "robonomist_title" for plot title.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.
#> roboplotr message: Using the attribute "source" for plot caption.
#> roboplotr warning: roboplotr attempts to make continous line with different patterns along 'Sarjatyyppi' over 'time. Be sure to arrange the data in proper order. For complex data you are better off handling the data manipulation outside of roboplot().
#> Error in map(split_d, function(g) {    tracetype <- unique(g$roboplot.plot.type)    hoverlab <- case_when(tracetype == "pie" ~ "label", any(c("horizontal",         "horizontalfill", "horizontalstack") %in% g$roboplot.plot.type) &         is.null(pattern) & quo_name(color) != ticktypes$y ~ "text",         TRUE ~ "customdata")    hovertemplate <- roboplotr_hovertemplate(hovertext, lab = hoverlab,         ticktypes)    legend_rank <- mean(g$roboplot.legend.rank)    if (tracetype == "pie") {        g <- mutate(g, roboplot.bg.color = roboplotr_alter_color(.data$roboplot.trace.color,             "dark"), roboplot.tx.color = roboplotr_text_color_picker(roboplotr_alter_color(.data$roboplot.trace.color,             "dark")), roboplot.in.tx.color = roboplotr_text_color_picker(.data$roboplot.trace.color))    }    if (is.null(highlight)) {        if (!is.null(pattern_showlegend)) {            show.legend <- pattern_showlegend[as.character(unique(g[[as_name(pattern)]]))]        }        else {            show.legend <- trace_showlegend        }    }    else {        show.legend <- roboplotr_highlight_legend(highlight,             g)    }    roboplotr_check_param(legend_title, c("logical", "character"),         allow_null = F)    legendgrouptitle <- if (legend_title == F) {        NULL    }    else if (is.character(legend_title)) {        str_c("<b>", legend_title, "</b>")    }    else {        str_c("<b>", as_name(color), "</b>")    }    if (!is.null(error_bars)) {        if (quo_is_null(error_bars$error_x)) {            error_x <- NULL        }        else {            error_x <- list(array = error_bars$error_x, color = error_bars$xcolor)        }        if (quo_is_null(error_bars$error_y)) {            error_y <- NULL        }        else {            error_y <- list(array = error_bars$error_y, color = error_bars$ycolor)        }    }    else {        error_x <- NULL        error_y <- NULL    }    plotting_params <- list(color = color, customdata = if (any(c("horizontal",         "horizontalfill", "horizontalstack") %in% g$roboplot.plot.mode)) {        ~roboplot.horizontal.label    } else {        ~roboplot.plot.text    }, data = g, direction = "clockwise", error_x = error_x,         error_y = error_y, hoverlabel = list(bgcolor = ~roboplot.bg.color,             color = ~roboplot.tx.color), hovertemplate = hovertemplate,         insidetextfont = list(family = getOption("roboplot.font.main")$family,             size = getOption("roboplot.font.main")$size, color = ~roboplot.in.tx.color),         labels = color, legendgroup = color, legendrank = legend_rank,         line = ~list(width = roboplot.linewidth, dash = roboplot.dash,             smoothing = 0.5, shape = case_when(roboplot.plot.mode ==                 "step" ~ "hv", roboplot.plot.mode == "smooth" ~                 "spline", TRUE ~ "linear")), marker = list(colors = ~roboplot.trace.color,             line = getOption("roboplot.trace.border"), symbol = markers$symbol,             size = markers$size, pattern = ~list(shape = roboplot.pattern)),         mode = str_replace_all(unique(g$roboplot.plot.mode),             c(`^smooth$` = "line", `^line$` = "lines", `^step$` = "lines",                 `^scatter$` = "markers", `scatter\\+line` = "markers+lines")),         name = ~if (!is.null(legend_maxwidth)) {            str_trunc(as.character(roboplot.plot.text), legend_maxwidth)        } else {            roboplot.plot.text        }, offset = if ("horizontalfill" %in% g$roboplot.plot.mode) {            ~roboplot.bar.offset        } else {            NULL        }, offsetgroup = ~str_c(roboplot.pattern, roboplot.trace.color),         orientation = ifelse(any(c("horizontal", "horizontalfill",             "horizontalstack") %in% g$roboplot.plot.mode) & "bar" %in%             g$roboplot.plot.type, "h", "v"), rotation = rotation,         showlegend = show.legend, sort = F, text = if (any(c("horizontal",             "horizontalfill", "horizontalstack") %in% g$roboplot.plot.mode)) {            ~roboplot.horizontal.label        } else {            ~roboplot.plot.text        }, textinfo = "percent", textposition = case_when("horizontal" %in%             g$roboplot.plot.mode ~ "auto", tracetype == "pie" ~             "inside", TRUE ~ "none"), texttemplate = if (tracetype ==             "pie") {            NULL        } else if ("horizontal" %in% g$roboplot.plot.mode) {            "%{x:,.1f}"        } else {            NA        }, type = ~tracetype, legendgrouptitle = list(text = legendgrouptitle,             font = getOption("roboplot.font.main")[c("color",                 "family", "size")]), values = as.formula(str_c("~",             ticktypes$y)), width = if ("horizontalfill" %in%             g$roboplot.plot.mode) {            ~roboplot.bar.width        } else {            NULL        }, x = as.formula(str_c("~", ticktypes$x)), y = as.formula(str_c("~",             ifelse(!is.null(legend_maxwidth) & ticktypes$y !=                 "value", "roboplot.trunc", ticktypes$y))))    shared_params <- c("data", "text", "texttemplate", "hovertemplate",         "legendgroup", "showlegend", "type", "legendgrouptitle",         "customdata", "error_y", "error_x")    if (tracetype %in% "scatter" & any(c("line", "step", "smooth") %in%         g$roboplot.plot.mode)) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "line", "mode", "name", "color", "xhoverformat")]    }    else if (tracetype == "scatter" & "scatter+line" %in% g$roboplot.plot.mode) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "line", "mode", "name", "marker", "color", "xhoverformat")]    }    else if (tracetype == "scatter" & "scatter" %in% g$roboplot.plot.mode) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "mode", "name", "color", "marker", "xhoverformat")]    }    else if (tracetype == "bar" & any(c("horizontal", "horizontalfill",         "horizontalstack") %in% g$roboplot.plot.mode)) {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "offsetgroup", "orientation", "offset", "width",             "color", "name", "textposition", "marker")]    }    else if (tracetype == "bar") {        plotting_params <- plotting_params[c(shared_params, "x",             "y", "offsetgroup", "name", "color", "textposition",             "marker", "xhoverformat")]    }    else if (tracetype == "pie") {        plotting_params <- plotting_params[c(shared_params, "labels",             "textposition", "textinfo", "insidetextfont", "direction",             "rotation", "sort", "hoverlabel", "marker", "values")]    }    if (!is.null(secondary_yaxis)) {        if (length(unique(as.numeric(pull(g, {            {                secondary_yaxis            }        })))) > 1) {            stop("The column referred to by argument 'secondary_yaxis' must be plotted to one axis per trace. Try setting it to match the argument 'color'.",                 call. = F)        }        if (unique(as.numeric(pull(g, {            {                secondary_yaxis            }        }))) == 2) {            plotting_params$yaxis <- "y2"        }    }    plotting_params}): ℹ In index: 1.
#> Caused by error in `if (tracetype == "pie") ...`:
#> ! missing value where TRUE/FALSE needed

# Finally, control the patterns for linetypes and bars with the parameter
# 'pattern_types', with a named vector containing either all the observations
# in the column 'pattern', or ".other" as a catch-all category.
energiantuonti |>
  dplyr::filter(Alue %in% c("Kanada", "Belgia", "Ruotsi"), Suunta == "Tuonti") |>
  roboplot(Alue, pattern = set_pattern(Alue, pattern_types = c(
    "Kanada" = "dash", ".other" = "dot"
  )))
#> roboplotr message: Using the attribute "robonomist_title" for plot title.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.
#> roboplotr message: Using the attribute "source" for plot caption.

# Bar plots use the pattern_types too, but they are different from the ones
# used by line plots. Don't worry if you don't give the corrent ones, roboplot()
#informs you which you should be using.
energiantuonti |>
  dplyr::filter(Alue %in% c("Kanada", "Belgia", "Ruotsi"), Suunta == "Tuonti") |>
  roboplot(Alue,
           plot_type = "bar",
           pattern = set_pattern(Alue, pattern_types = c(
    "Ruotsi" = "", ".other" = "x"
  )))
#> roboplotr message: Using the attribute "robonomist_title" for plot title.
#> roboplotr message: No acceptable frequency attribute was detected for hoverlabel from data 'd', and has not been provided as "frequency" in the argument 'hovertext'.
#> Guessing frequency "Quarterly" for date format %YQ%q.
#> roboplotr message: Using the attribute "source" for plot caption.