Set global options for roboplot()

Override the default options for colors, fonts etc. for any plot created with roboplot()

Usage

set_roboplot_options(
  accessible = NULL,
  artefacts = NULL,
  border = NULL,
  background_color = NULL,
  caption_template = NULL,
  dashtypes = NULL,
  font_main = NULL,
  font_title = NULL,
  font_caption = NULL,
  grid = NULL,
  height = NULL,
  imgdl_wide = NULL,
  imgdl_narrow = NULL,
  imgdl_small = NULL,
  infobox = NULL,
  linewidth = NULL,
  locale = NULL,
  logo_file = NULL,
  markers = NULL,
  modebar = NULL,
  patterns = NULL,
  tick_colors = NULL,
  trace_border = NULL,
  trace_colors = NULL,
  xaxis_ceiling = NULL,
  zeroline = NULL,
  verbose = NULL,
  width = NULL,
  shinyapp = NULL,
  reset = F
)

Arguments

accessible

Logical. Forces trace colors accessible against plot background.

artefacts

Function. Control html and other file creation. Use set_artefacts().

border

Function. Control plot borders. Use set_border().

background_color

Character. Plot background color. Must be a hexadecimal color or a valid css color.

caption_template

Character. Template for str_glue() used to parse captions.

dashtypes

Character vector. Line trace linetypes in order of usage. Must contain all of "solid", "dash", "dot", "longdash", "dashdot", and "longdashdot" in any order.

font_main, font_title, font_caption

Functions. Use set_font().

grid

Function. Control the grid behind the traces. Use set_grid().

height, width

Numerics. Height and width of roboplotr plots in pixels. Use NA for viewport size.

imgdl_wide, imgdl_narrow, imgdl_small

Functions. Use set_imgdl_layout. Controls the dimensions and fonts of image files downloaded through modebar buttons.

infobox

Function. Defines the appearance of the infobox available from modebar when 'info_text' is provided for roboplot() or robotable(). Use set_infobox().

linewidth

Numeric. The default roboplotr line trace width.

locale

Function. Defines locale parameters as roboplot() needs them. Use set_locale().

logo_file

Character. The filepath to the logo used in every plot.

markers

Function. Control marker appearance. Use set_markers().

modebar

Character vector. Buttons contained in modebar in the given order. Must contain any of "home", "closest", "compare", "zoomin", "zoomout", "img_w", "img_n", "img_s", "data_dl" and "robonomist" in any order.

patterns

Character vector. Line trace linetypes in order of usage. Must contain all of "", "/", "\", "x", "-", "|", "+" and "." in any order.

tick_colors

List. Plot tick element colors. Values need to be hexadecimal colors or valid css colors, named "x" and "y".

trace_border

List. Borders for bars, pies and markers. Values must be named "color" and "width". Item "color" needs to be a hexadecimal color or a valid css color, item "width" needs to be numeric.

trace_colors

Character vector. Trace colors in order of usage. Needs to be a hexadecimal color or a valid css color. You should provide enough colors for most use cases, while roboplotr adds colors as needed.

xaxis_ceiling

Character. Default rounding for yaxis limit. One of "default", "days", "months", "weeks", "quarters", "years" or "guess".

zeroline

Function. Control the appearance of zeroline when set using roboplot() parameter 'zeroline'. Use set_zeroline().

verbose

Character. Will roboplot display all messages, alerts and warnings, or warnings only? Must be one of "All", "Alert", or "Warning".

shinyapp

Logical or list. Makes fonts, css and javascript available for shiny apps. If given as list, use a named list with the single character string named "container" describing the css selector for the element in a shiny app where most roboplots will be contained in. Used for relayouts if the plot is rendered while the container is not displayed.

reset

Logical. Ignores other options, resets options to defaults.

Examples